refactor : 기존 aboutsopt 모듈 및 패키지 삭제 + Dto들만  이전
