+++ src/main/java/sopt/org/homepage/main/dto/response/GetAboutPageResponseDto.java
@@ -6,7 +6,6 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.validation.annotation.Validated;
-import sopt.org.homepage.aboutsopt.dto.GetAboutSoptResponseDto;
 import sopt.org.homepage.main.dto.response.main.branding.GetMainBrandingColorResponseRecordDto;
 import sopt.org.homepage.main.dto.response.main.core.GetMainCoreValueResponseRecordDto;
 import sopt.org.homepage.main.dto.response.main.curriculum.GetMainPartCurriculumResponseRecordDto;
